{"schema":"decodex/commit/1","summary":"Split search API route handlers","authority":"manual"}
